Limit the number of bot tensors applied to each bot to 50.
This drastically improves both performance and gameplay.

Previously a swarm of >100 bots would be impossible to control since
they have such a strong gravitational core. Now that core is limited,
you can play with it. It feels like fluid. It's incredible.

A shame I didn't think of doing this long ago, but the competition has
been extended by 1 hour so I am going to sneak this change in.