Add dynamic tensor update system.
Causes bots to only do tensor updates every N ticks.

Currently just set to do it 100% of the time.

If you turn this on, it's still a problem since you get a tensor update every N
ticks and a massive lag spike on that frame. Instead, we need to spread it out
in a round-robin style.