Set web thread mode to single threaded.
This allows for much simpler server config (can be run from any static web
host).

Set music playback type to "default". On the web, this will be Sample (instead
of Stream which is what it was set to; not sure why), which has much better
latency in single-threaded. Without this, the music would stutter awfully.